Conflicts:
	src/server/controllers/LibrarianController.java

+++ src/client/controllers/ManageStockClientController.java
@@ -23,6 +23,7 @@ import javafx.scene.control.ListView;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.TextField;
+import javafx.scene.control.Alert.AlertType;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.Pane;
@@ -198,10 +199,30 @@ public class ManageStockClientController {
     @FXML
     void onClickAddCopy(ActionEvent event) 
     {
+    	tfEnterNewCopyID.setStyle(null);
+    	if(tfEnterNewCopyID.getText().isEmpty())
+    		{	tfEnterNewCopyID.setStyle("-fx-border-color: red ; -fx-border-width: 2px ;-fx-border-radius: 5px;");
+    			return;
+    		}
+    	Book selectedBook=ViewStarter.client.searchBookOnManageStockControllerObj.getBooksTable().getSelectionModel().getSelectedItem();
+    	if(selectedBook==null)
+    		{Utils utils = new Utils(ViewStarter.client.mainViewController);
+    		utils.showAlertWithHeaderText(AlertType.ERROR, "", "No book selected! please select book");
+    		return;
+    		}
+    	
     	Object[] queryArr=new Object[3];
-    	int catlogNumOfSelectedBook=ViewStarter.client.searchBookOnManageStockControllerObj.getBooksTable().getSelectionModel().getSelectedItem().getCatalogNum();
-    	int copiesNumOfSelectedBook=ViewStarter.client.searchBookOnManageStockControllerObj.getBooksTable().getSelectionModel().getSelectedItem().getCopiesNum();
-    	int avilableCopiesNumOfSelectedBook=ViewStarter.client.searchBookOnManageStockControllerObj.getBooksTable().getSelectionModel().getSelectedItem().getAvilableCopiesNum();
+    	
+    	int copiesNumOfSelectedBook=selectedBook.getCopiesNum();
+    	int avilableCopiesNumOfSelectedBook=selectedBook.getAvilableCopiesNum();
+    	int catlogNumOfSelectedBook=selectedBook.getCatalogNum();
+    	for (Copy cp: tvCopies.getItems())
+    		if (cp.getCopyID().equals(tfEnterNewCopyID.getText()))
+    				{
+    					Utils utils = new Utils(ViewStarter.client.mainViewController);
+    					utils.showAlertWithHeaderText(AlertType.ERROR, "", "Copy already exists!");
+    					return;
+    				}
     	queryArr[0]="INSERT INTO obl.copeis (copyID,bCatalogNum, isAvilable) Values('"+tfEnterNewCopyID.getText()+"',"+catlogNumOfSelectedBook+",1);";
     	queryArr[1]="UPDATE obl.books SET bCopiesNum="+(copiesNumOfSelectedBook+1)+",bAvilableCopiesNum="+(avilableCopiesNumOfSelectedBook+1)+" WHERE bCatalogNum="+catlogNumOfSelectedBook+";";
     	queryArr[2]= new Copy(tfEnterNewCopyID.getText(),catlogNumOfSelectedBook,true);
@@ -260,10 +281,24 @@ public class ManageStockClientController {
     @FXML
     void onClickDeleteCopy(ActionEvent event)
     {
-    	String copyIDofSelected=tvCopies.getSelectionModel().getSelectedItem().getCopyID();
-    	int catlogNumOfSelectedBook=tvCopies.getSelectionModel().getSelectedItem().getbCatalogNum();
-      	int copiesNumOfSelectedBook=ViewStarter.client.searchBookOnManageStockControllerObj.getBooksTable().getSelectionModel().getSelectedItem().getCopiesNum();
-    	int avilableCopiesNumOfSelectedBook=ViewStarter.client.searchBookOnManageStockControllerObj.getBooksTable().getSelectionModel().getSelectedItem().getAvilableCopiesNum();
+    	Copy selectedCopy=tvCopies.getSelectionModel().getSelectedItem();
+    	if(selectedCopy==null)
+    	{
+    		Utils utils = new Utils(ViewStarter.client.mainViewController);
+    		utils.showAlertWithHeaderText(AlertType.ERROR, "", "No copy selected! please select copy");
+    		return;	
+    	}
+    	String copyIDofSelected=selectedCopy.getCopyID();
+    	Book selectedBook=ViewStarter.client.searchBookOnManageStockControllerObj.getBooksTable().getSelectionModel().getSelectedItem();
+     	if(selectedBook==null)
+    	{
+    		Utils utils = new Utils(ViewStarter.client.mainViewController);
+    		utils.showAlertWithHeaderText(AlertType.ERROR, "", "No book selected! please select book");
+    		return;	
+    	}
+    	int catlogNumOfSelectedBook=selectedCopy.getbCatalogNum();
+      	int copiesNumOfSelectedBook=selectedBook.getCopiesNum();
+    	int avilableCopiesNumOfSelectedBook=selectedBook.getAvilableCopiesNum();
     	String[] copyArr= new String[2];
     	copyArr[0]="Delete from obl.copeis where copyID='"+copyIDofSelected+"';";
     	copyArr[1]="UPDATE obl.books SET bCopiesNum="+(copiesNumOfSelectedBook-1)+",bAvilableCopiesNum="+(avilableCopiesNumOfSelectedBook-1)+" WHERE bCatalogNum="+catlogNumOfSelectedBook+";";
